+++ admisiones/forms/admisiones_forms.py
@@ -29,6 +29,10 @@ class InformeTecnicoJuridicoForm(forms.ModelForm):
             "organizacion_avalista_2",
             "material_difusion_vinculado",
         ]
+        labels = {
+            "IF_relevamiento_territorial": "IF Relevamiento Territorial",
+            "if_relevamiento": "IF Relevamiento",
+        }
 
     def __init__(self, *args, **kwargs):
         admision = kwargs.pop("admision", None)
@@ -53,10 +57,20 @@ class InformeTecnicoJuridicoForm(forms.ModelForm):
                 anexo = Anexo.objects.get(admision=admision)
                 nombre = anexo.responsable_nombre or ""
                 apellido = anexo.responsable_apellido or ""
+                comedor = admision.comedor
+                organizacion = comedor.organizacion
 
                 self.fields["expediente_nro"].initial = anexo.expediente
+                self.fields["nombre_organizacion"].initial = organizacion.nombre
+                self.fields["cuit_organizacion"].initial = organizacion.cuit
+                self.fields["mail_organizacion"].initial = organizacion.email
                 self.fields["nombre_espacio"].initial = anexo.efector
                 self.fields["tipo_espacio"].initial = anexo.tipo_espacio
+                self.fields["tipo_espacio"].initial = anexo.tipo_espacio
+                self.fields["barrio_espacio"].initial = comedor.barrio
+                self.fields["localidad_espacio"].initial = comedor.localidad
+                self.fields["partido_espacio"].initial = comedor.partido
+                self.fields["provincia_espacio"].initial = comedor.provincia
                 self.fields["domicilio_espacio"].initial = anexo.domicilio
                 self.fields["responsable_tarjeta_nombre"].initial = (
                     f"{nombre} {apellido}".strip()
@@ -104,10 +118,20 @@ class InformeTecnicoBaseForm(forms.ModelForm):
                 anexo = Anexo.objects.get(admision=admision)
                 nombre = anexo.responsable_nombre or ""
                 apellido = anexo.responsable_apellido or ""
+                comedor = admision.comedor
+                organizacion = comedor.organizacion
 
                 self.fields["expediente_nro"].initial = anexo.expediente
+                self.fields["nombre_organizacion"].initial = organizacion.nombre
+                self.fields["cuit_organizacion"].initial = organizacion.cuit
+                self.fields["mail_organizacion"].initial = organizacion.email
                 self.fields["nombre_espacio"].initial = anexo.efector
                 self.fields["tipo_espacio"].initial = anexo.tipo_espacio
+                self.fields["tipo_espacio"].initial = anexo.tipo_espacio
+                self.fields["barrio_espacio"].initial = comedor.barrio
+                self.fields["localidad_espacio"].initial = comedor.localidad
+                self.fields["partido_espacio"].initial = comedor.partido
+                self.fields["provincia_espacio"].initial = comedor.provincia
                 self.fields["domicilio_espacio"].initial = anexo.domicilio
                 self.fields["responsable_tarjeta_nombre"].initial = (
                     f"{nombre} {apellido}".strip()
@@ -340,3 +364,4 @@ class AnexoForm(forms.ModelForm):
 
             self.fields["efector"].initial = comedor.nombre
             self.fields["domicilio"].initial = f"{calle} {numero}".strip()
+            self.fields["expediente"].initial = admision.num_expediente
