+++ admisiones/templates/admisiones_legales_detalle.html
@@ -14,7 +14,7 @@
 {% block breadcrumb %}
     <ol class="breadcrumb float-sm-right">
         <div class="breadcrumb-item">
-            <a href="{% url 'admisiones_legales_listar' %}">Legales - Expedientes</a>
+            <a href="{% url 'admisiones_legales_listar' %}">Expedientes</a>
         </div>
         <div class="breadcrumb-item active">{{ admision.num_expediente }}</div>
     </ol>
@@ -70,17 +70,12 @@
                 <button type="button"
                         class="btn btn-primary mr-1"
                         data-bs-toggle="modal"
-                        data-bs-tipo="Resolución"
-                        data-bs-label="Cargue el Número de Resolución"
-                        data-bs-target="#modalDocumentoExpediente">Resolución</button>
-            {% elif admision.estado_legales == "Resolución Generada" %}
-                <button type="button"
-                        class="btn btn-primary mr-1"
-                        data-bs-toggle="modal"
-                        data-bs-tipo="Firma Convenio"
-                        data-bs-label="Cargue el IF del Convenio"
-                        data-bs-target="#modalDocumentoExpediente">Firma Convenio</button>
-            {% elif admision.estado_legales == "Convenio Firmado" %}
+                        data-bs-tipo="Disposición"
+                        data-bs-label="Cargue el Número de Disposición"
+                        data-bs-target="#modalDocumentoExpediente">Disposición</button>
+            {% elif admision.estado_legales == "Disposición Generada" %}
+                {% comment %}<button type="button" class="btn btn-primary mr-1" data-bs-toggle="modal" data-bs-tipo="Firma Convenio" data-bs-label="Cargue el IF del Convenio" data-bs-target="#modalDocumentoExpediente">Firma Convenio</button>
+            {% elif admision.estado_legales == "Convenio Firmado" %}{% endcomment %}
                 <button type="button"
                         class="btn btn-primary mr-1"
                         data-bs-toggle="modal"
@@ -142,14 +137,15 @@
                         </li>
                         <li class="nav-item">
                             <div class="comment-text">
-                                Resolución:
-                                {% if value_resolucion %}
-                                    <span class="fw-bold">{{ value_resolucion }}</span> <i class="bi bi-check-circle-fill text-success"></i>
+                                Disposición:
+                                {% if value_disposicion %}
+                                    <span class="fw-bold">{{ value_disposicion }}</span> <i class="bi bi-check-circle-fill text-success"></i>
                                 {% else %}
                                     <i class="bi bi-x-circle-fill text-danger"></i>
                                 {% endif %}
                             </div>
                         </li>
+                        {% comment %}
                         <li class="nav-item">
                             <div class="comment-text">
                                 Firma Convenio:
@@ -160,6 +156,7 @@
                                 {% endif %}
                             </div>
                         </li>
+                        {% endcomment %}
                         <li class="nav-item">
                             <div class="comment-text">
                                 Numero CONV:
@@ -210,6 +207,17 @@
                                 <div class="comment-text">
                                     Informe Técnico:
                                     <a href="{{ pdf_url }}" target="_blank"><i class="bi bi-filetype-pdf text-info h4 ms-3 text-bold"></i></a>
+                                </div>
+                            </li>
+                        {% endif %}
+                        {% if informes_complementarios %}
+                            <li class="nav-item">
+                                <div class="comment-text">
+                                    Informes Complementarios:
+                                    {% for i in informes_complementarios %}
+                                        <span class="fw-bold"><a href="{{ i.pdf.url }}" target="_blank"><i class="bi bi-filetype-pdf text-info h4 ms-3 text-bold"></i>{{ i.creado |date:'d/m/Y' }}</a></span>
+                                        <br />
+                                    {% endfor %}
                                 </div>
                             </li>
                         {% endif %}
@@ -321,7 +329,15 @@
                                     <tr id="fila-{{ doc.id }}">
                                         <td class="ps-3">{{ cambio.fecha|date:'d/m/Y H:i' }}</td>
                                         <td class="ps-3">{{ cambio.campo }}</td>
-                                        <td class="ps-3">{{ cambio.valor_nuevo| default_if_none:"-" }}</td>
+                                        <td class="ps-3">
+                                            {% if cambio.valor_nuevo == "True" %}
+                                                Si
+                                            {% elif cambio.valor_nuevo == "False" %}
+                                                No
+                                            {% else %}
+                                                {{ cambio.valor_nuevo| default_if_none:"-" }}
+                                            {% endif %}
+                                        </td>
                                         <td class="ps-3">{{ cambio.usuario.username }}</td>
                                     </tr>
                                 {% endfor %}
@@ -407,7 +423,7 @@
                 <form method="POST" action="">
                     {% csrf_token %}
                     <div class="modal-header">
-                        <h5 class="modal-title">Formulario RESO</h5>
+                        <h5 class="modal-title">Formulario Disposición</h5>
                         <button type="button"
                                 class="btn-close"
                                 data-bs-dismiss="modal"
