+++ admisiones/templates/informe_tecnico_complementario_detalle.html
@@ -1,0 +1,83 @@
+{% extends "includes/main.html" %}
+{% load static %}
+{% load static custom_filters %}
+{% block title %}Informe Complementario: {{ admision.comedor.nombre }}{% endblock %}
+{% block titulo-pagina %}Informe Complementario: {{ admision.comedor.nombre }}{% endblock %}
+{% block breadcrumb %}
+    <ol class="breadcrumb float-sm-right">
+        <div class="breadcrumb-item">
+            <a href="{% url 'admisiones_tecnicos_editar' admision.id %}">{{ admision.comedor.nombre }}</a>
+        </div>
+        <div class="breadcrumb-item active">Informe Complementario</div>
+    </ol>
+{% endblock %}
+{% block content %}
+    <div class="container mt-4">
+        <div class="row justify-content-between d-print-none mx-1 mb-3">
+            <div class="col-auto">
+                <a href="{% url 'admisiones_tecnicos_editar' admision.id %}"
+                   class="btn btn-secondary mr-1 backbutton-handler">Volver</a>
+            </div>
+        </div>
+        <div class="row">
+            <form method="post">
+                {% csrf_token %}
+                <div class="list-group">
+                    {% for nombre, valor in campos %}
+                        <div class="list-group-item d-flex flex-column gap-2">
+                            <div class="d-flex justify-content-between align-items-center">
+                                <div>
+                                    <h6 class="mb-1">{{ nombre }}</h6>
+                                    <p class="mb-1">{{ valor }}</p>
+                                </div>
+                                <div class="form-check">
+                                    <input class="form-check-input campo-subsanar"
+                                           type="checkbox"
+                                           name="campos_subsanar"
+                                           value="{{ nombre }}"
+                                           id="check_{{ forloop.counter }}" />
+                                    <label class="form-check-label" for="check_{{ forloop.counter }}">Modificar campo</label>
+                                </div>
+                            </div>
+                            <div class="input-wrapper mt-2"
+                                 id="input_{{ forloop.counter }}"
+                                 style="display: none">
+                                <label for="campo_{{ forloop.counter }}">Nuevo valor para {{ nombre }}:</label>
+                                <input type="text"
+                                       class="form-control"
+                                       name="campo_{{ nombre }}"
+                                       id="campo_{{ forloop.counter }}"
+                                       placeholder="Escriba el nuevo valor" />
+                            </div>
+                        </div>
+                    {% endfor %}
+                </div>
+                <div class="mt-4">
+                    <a href="{% url 'comedor_detalle' admision.comedor_id %}"
+                       class="btn btn-secondary">Volver</a>
+                    <button type="submit" class="btn btn-primary">Guardar</button>
+                </div>
+            </form>
+        </div>
+    </div>
+{% endblock %}
+{% block customJS %}
+    <script>
+  document.querySelectorAll(".campo-subsanar").forEach(function (checkbox) {
+    checkbox.addEventListener("change", function () {
+      const id = this.id.replace("check_", "input_");
+      const inputWrapper = document.getElementById(id);
+      const inputField = inputWrapper?.querySelector("input");
+
+      if (this.checked) {
+        inputWrapper.style.display = "block";
+        inputField.required = true;
+      } else {
+        inputWrapper.style.display = "none";
+        inputField.value = "";
+        inputField.required = false;
+      }
+    });
+  });
+    </script>
+{% endblock %}
